Updated PA-2 Source Code
This updated code has implemented the merge sort portion of the assignment. It generates nine vectors (arrays) of size n * 1000, populates the elements with values between -16383 and 16384, and allows the user to input a number (1-9) to output both the original and the sorted versions of the vector. 

TODO: The output needs to be cleaned up to be more user-readable. Additionally, the time-tracking code has not been implemented, as well as the automatic excel generation. The code comments should also be cleaned up for the final production.